# Part2/swagger/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: '0.1.0'
  title: Engine API
  description: An API to gamify applications
host: localhost:8080
basePath: /api
schemes:
  - http
paths:
  /badges:
    post:
      description: create a badge
      operationId: createBadge
      consumes:
        - application/json
      parameters:
        - name: badge
          in: body
          required: true
          schema:
            $ref: '#/definitions/BadgeWithoutId'
      responses:
        '201':
          description: created
          schema:
            type: object
    get:
      description: get the list of all badges
      operationId: getBadges
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Badge'
  /badges/{badgeId}:
    get:
      description: Gets a badge by ID.
      operationId: getBadgesById
      produces:
        - application/json
      parameters:
        - in: path
          name: badgeId
          type: integer
          format: int64
          required: true
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Badge'
  /pointScales:
    post:
      description: create a pointScale
      operationId: createPointScale
      consumes:
        - application/json
      parameters:
        - name: pointScale
          in: body
          required: true
          schema:
            $ref: '#/definitions/PointScaleWithoutId'
      responses:
        '201':
          description: created
          schema:
            type: object
    get:
      description: get the list of all pointScales
      operationId: getPointScales
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/PointScale'
  /pointScales/{pointScaleId}:
    get:
      description: Gets a PointScale by ID.
      operationId: getPointScaleById
      produces:
        - application/json
      parameters:
        - in: path
          name: pointScaleId
          type: integer
          format: int64
          required: true
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/PointScale'
  /events:
    post:
      description: send an event
      operationId: sendEvent
      consumes:
        - application/json
      parameters:
        - name: event
          in: body
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
        '201':
          description: created
          schema:
            type: object
  /rules:
    post:
      description: create a rule
      operationId: createRule
      consumes:
        - application/json
      parameters:
        - name: rule
          in: body
          required: true
          schema:
            $ref: '#/definitions/RuleWithoutId'
      responses:
        '201':
          description: created
          schema:
            type: object
    get:
      description: get the list of all rules
      operationId: getRules
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Rule'
  /rules/{ruleId}:
    get:
      description: Gets a PointScale by ID.
      operationId: getRuleById
      produces:
        - application/json
      parameters:
        - in: path
          name: ruleId
          type: integer
          format: int64
          required: true
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Rule'

definitions:
  Rule:
    allOf:
      - properties:
          id:
            type: integer
            format: int64
      - $ref: '#/definitions/RuleWithoutId'
  RuleWithoutId:
    type: object
    properties:
      if:
        type: object
        properties:
          type:
            type: string
      then:
        type: object
        properties:
          awardBadgeId:
            type: integer
            format: int64
          awardPoints:
            type: object
            properties:
              pointScaleId:
                type: integer
                format: int64
              amount:
                type: integer
  Badge:
    allOf:
      - properties:
          id:
            type: integer
            format: int64
      - $ref: '#/definitions/BadgeWithoutId'
  BadgeWithoutId:
    type: object
    properties:
      name:
        type: string
      image:
        type: string
  PointScale:
    allOf:
      - properties:
          id:
            type: integer
            format: int64
      - $ref: '#/definitions/PointScaleWithoutId'
  PointScaleWithoutId:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
  Event:
    type: object
    properties:
      userId:
        type: string
      type:
        type: string
      timestamp:
        type: string
        format: date-time